+++ libs/base_core/src/text/char/impls/traits.rs
@@ -3,8 +3,10 @@
 //! implementations of core traits
 //
 
-use crate::{DataOverflow, paste, text::char::*, unwrap};
-use ::core::fmt;
+use crate::{
+    Binary, DataOverflow, Debug, Display, FmtResult, Formatter, LowerHex, Octal, UpperHex, paste,
+    text::char::*, unwrap,
+};
 
 /* Default, Display, Debug */
 
@@ -15,7 +17,7 @@ macro_rules! char_core_impls {
             char7 + Self(unwrap![some NonExtremeU8::new(0)]),
             char8 + Self(0),
             char16 + Self(unwrap![some NonSurrogateU16::new(0)]),
-            // char_utf8 + Self(unwrap![some NonSurrogateU32::new(0)]) // TODO
+            char_utf8 + Self(unwrap![some NonExtremeU32::new(0)])
         ];
     };
     ($( $name:ident + $default:expr ),+ $(,)?) => {
@@ -26,39 +28,34 @@ macro_rules! char_core_impls {
             /// Returns the default value of `\x00` (nul character).
             fn default() -> Self { $default }
         }
-        // TEMP
-        // impl ConstDefault for super::$name {
-        //     /// Returns the default value of `\x00` (nul character).
-        //     const DEFAULT: Self = $default;
-        // }
-        impl fmt::Display for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        impl Display for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
                 write!(f, "{}", self.to_char())
             }
         }
-        impl fmt::Debug for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        impl Debug for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
                 write!(f, "{:?}", self.to_char())
             }
         }
-        impl fmt::Binary for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                fmt::Binary::fmt(&self.to_scalar(), f)
+        impl Binary for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
+                Binary::fmt(&self.to_scalar(), f)
             }
         }
-        impl fmt::LowerHex for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                fmt::LowerHex::fmt(&self.to_scalar(), f)
+        impl LowerHex for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
+                LowerHex::fmt(&self.to_scalar(), f)
             }
         }
-        impl fmt::UpperHex for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                fmt::UpperHex::fmt(&self.to_scalar(), f)
+        impl UpperHex for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
+                UpperHex::fmt(&self.to_scalar(), f)
             }
         }
-        impl fmt::Octal for super::$name {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                fmt::Octal::fmt(&self.to_scalar(), f)
+        impl Octal for super::$name {
+            fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult<()> {
+                Octal::fmt(&self.to_scalar(), f)
             }
         }
     }};
@@ -66,92 +63,75 @@ macro_rules! char_core_impls {
 #[rustfmt::skip]
 char_core_impls!();
 
-/* From char7 */
-
+/// From char7
+#[rustfmt::skip]
 mod c7 {
-    use super::super::*;
-
-    impl From<char7> for char {
-        fn from(c: char7) -> char {
-            c.to_char()
-        }
-    }
-    impl From<char7> for char8 {
-        fn from(c: char7) -> char8 {
-            c.to_char8()
-        }
-    }
-    impl From<char7> for char16 {
-        fn from(c: char7) -> char16 {
-            c.to_char16()
-        }
-    }
+    use super::*;
+    impl From<char7> for char { fn from(c: char7) -> char { c.to_char() } }
+    impl From<char7> for char8 { fn from(c: char7) -> char8 { c.to_char8() } }
+    impl From<char7> for char16 { fn from(c: char7) -> char16 { c.to_char16() } }
+    impl From<char7> for char_utf8 { fn from(c: char7) -> char_utf8 { c.to_char_utf8() } }
 }
-
-/* From char8 */
-
+/// From char8
+#[rustfmt::skip]
 mod c8 {
     use super::*;
-
-    impl From<char8> for char {
-        fn from(c: char8) -> char {
-            c.to_char()
-        }
-    }
-    impl TryFrom<char8> for char7 {
-        type Error = DataOverflow;
-        fn try_from(c: char8) -> Result<char7, DataOverflow> {
-            c.try_to_char7()
-        }
-    }
-    impl From<char8> for char16 {
-        fn from(c: char8) -> char16 {
-            c.to_char16()
-        }
+    impl From<char8> for char { fn from(c: char8) -> char { c.to_char() } }
+    impl From<char8> for char16 { fn from(c: char8) -> char16 { c.to_char16() } }
+    impl From<char8> for char_utf8 { fn from(c: char8) -> char_utf8 { c.to_char_utf8() } }
+    impl TryFrom<char8> for char7 { type Error = DataOverflow;
+        fn try_from(c: char8) -> Result<char7, DataOverflow> { c.try_to_char7() }
     }
 }
-
-/* From char16 */
-
+/// From char16
+#[rustfmt::skip]
 mod c16 {
     use super::*;
-
-    impl From<char16> for char {
-        fn from(c: char16) -> char {
-            c.to_char()
-        }
-    }
+    impl From<char16> for char { fn from(c: char16) -> char { c.to_char() } }
+    impl From<char16> for char_utf8 { fn from(c: char16) -> char_utf8 { c.to_char_utf8() } }
     impl TryFrom<char16> for char7 {
         type Error = DataOverflow;
-        fn try_from(c: char16) -> Result<char7, DataOverflow> {
-            c.try_to_char7()
-        }
+        fn try_from(c: char16) -> Result<char7, DataOverflow> { c.try_to_char7() }
     }
     impl TryFrom<char16> for char8 {
         type Error = DataOverflow;
-        fn try_from(c: char16) -> Result<char8, DataOverflow> {
-            c.try_to_char8()
-        }
+        fn try_from(c: char16) -> Result<char8, DataOverflow> { c.try_to_char8() }
+    }
+}
+/// From char_utf8
+#[rustfmt::skip]
+mod utf8 {
+    use super::*;
+    impl From<char_utf8> for char { fn from(c: char_utf8) -> char { c.to_char() } }
+    impl TryFrom<char_utf8> for char7 {
+        type Error = DataOverflow;
+        fn try_from(c: char_utf8) -> Result<char7, DataOverflow> { c.try_to_char7() }
+    }
+    impl TryFrom<char_utf8> for char8 {
+        type Error = DataOverflow;
+        fn try_from(c: char_utf8) -> Result<char8, DataOverflow> { c.try_to_char8() }
+    }
+    impl TryFrom<char_utf8> for char16 {
+        type Error = DataOverflow;
+        fn try_from(c: char_utf8) -> Result<char16, DataOverflow> { c.try_to_char16() }
     }
 }
 
-/* From char */
-
-impl TryFrom<char> for char7 {
-    type Error = DataOverflow;
-    fn try_from(c: char) -> Result<char7, DataOverflow> {
-        char7::try_from_char(c)
+/// From char
+#[rustfmt::skip]
+mod c {
+    use super::*;
+    impl From<char> for char_utf8 { fn from(c: char) -> char_utf8 { char_utf8::from_char(c) } }
+    impl TryFrom<char> for char7 {
+        type Error = DataOverflow;
+        fn try_from(c: char) -> Result<char7, DataOverflow> { char7::try_from_char(c) }
     }
-}
-impl TryFrom<char> for char8 {
-    type Error = DataOverflow;
-    fn try_from(c: char) -> Result<char8, DataOverflow> {
-        char8::try_from_char(c)
+    impl TryFrom<char> for char8 {
+        type Error = DataOverflow;
+        fn try_from(c: char) -> Result<char8, DataOverflow> { char8::try_from_char(c) }
     }
-}
-impl TryFrom<char> for char16 {
-    type Error = DataOverflow;
-    fn try_from(c: char) -> Result<char16, DataOverflow> {
-        char16::try_from_char(c)
+    impl TryFrom<char> for char16 {
+        type Error = DataOverflow;
+        fn try_from(c: char) -> Result<char16, DataOverflow> { char16::try_from_char(c) }
     }
 }